requirements.txt
# Azure Core サービス
azure-identity==1.17.0

# Azure Functions
azure-functions==1.19.0

# Web フレームワーク
Flask[async]==3.0.3

# OpenAI (grpcio依存関係注意)
openai==1.55.3

# Azure サービス
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0

# 基本ライブラリ
python-dotenv==1.0.0
azure-cosmos==4.5.0
quart==0.19.9
uvicorn[standard]==0.24.0
aiohttp==3.10.11

# ASGIサーバー（Quartアプリ用・必須）
gunicorn==20.1.0
hypercorn==0.17.3
pydantic==2.8.2
pydantic-settings==2.2.1
azure-core>=1.35.0
# AI エージェント (grpcio依存なし確認済み)
azure-ai-agents>=1.0.2
azure-keyvault-secrets==4.8.0
requests==2.32.3
